Problem Set 6 --> Sentimental Hello
Problem to Solve

In a file called hello.py in a folder called sentimental-hello, implement a program that prompts a user for their name, and then prints hello, so-and-so, where so-and-so is their provided name, exactly as you did in Problem Set 1. Except that your program this time should be written in Python!