Problem Set 8 --> Homepage
Build a simple homepage using HTML, CSS, and JavaScript.

Background
The internet has enabled incredible things: we can use a search engine to research anything imaginable, communicate with friends and family members around the globe, play games, take courses, and so much more. But it turns out that nearly all pages we may visit are built on three core languages, each of which serves a slightly different purpose:

HTML, or HyperText Markup Language, which is used to describe the content of websites;
CSS, Cascading Style Sheets, which is used to describe the aesthetics of websites; and
JavaScript, which is used to make websites interactive and dynamic.
Create a simple homepage that introduces yourself, your favorite hobby or extracurricular, or anything else of interest to you.